+++ src/features/drawer/Routes.jsx
@@ -1,6 +1,9 @@
 // @ts-check
 import * as React from 'react'
 import ListItem from '@mui/material/ListItem'
+import ListItemText from '@mui/material/ListItemText'
+import Switch from '@mui/material/Switch'
+import { useTranslation } from 'react-i18next'
 
 import { useMemory } from '@store/useMemory'
 import { useStorage, useDeepStore } from '@store/useStorage'
@@ -9,8 +12,13 @@ import { SliderTile } from '@components/inputs/SliderTile'
 import { CollapsibleItem } from './components/CollapsibleItem'
 
 const RouteSlider = () => {
+  const { t } = useTranslation()
   const enabled = useStorage((s) => !!s.filters?.routes?.enabled)
   const [filters, setFilters] = useDeepStore('filters.routes.distance')
+  const [compactView, setCompactView] = useDeepStore(
+    'userSettings.routes.compactView',
+    true,
+  )
   const baseDistance = useMemory.getState().filters?.routes?.distance
 
   /** @type {import('@rm/types').RMSlider} */
@@ -31,6 +39,17 @@ const RouteSlider = () => {
 
   return (
     <CollapsibleItem open={enabled}>
+      <ListItem
+        secondaryAction={
+          <Switch
+            color="secondary"
+            onChange={(_, checked) => setCompactView(checked)}
+            checked={compactView !== false}
+          />
+        }
+      >
+        <ListItemText primary={t('compact_route_view')} />
+      </ListItem>
       <ListItem>
         <SliderTile
           slide={slider}