+++ src/features/route/index.js
@@ -3,3 +3,6 @@
 export * from './routeMarker'
 export * from './RoutePopup'
 export * from './RouteTile'
+export * from './RouteLayer'
+export * from './useRouteStore'
+export * from './constants'