+++ src/features/station/StationPopup.jsx
@@ -17,8 +17,6 @@ import Rating from '@mui/material/Rating'
 import Typography from '@mui/material/Typography'
 import Stack from '@mui/material/Stack'
 import Box from '@mui/material/Box'
-import LockOpenIcon from '@mui/icons-material/LockOpen'
-import LockIcon from '@mui/icons-material/Lock'
 
 import { useMemory } from '@store/useMemory'
 import { setDeepStore, useGetDeepStore, useStorage } from '@store/useStorage'
@@ -37,7 +35,7 @@ import {
   ExpandWithState,
 } from '@components/inputs/ExpandCollapse'
 import { VirtualGrid } from '@components/virtual/VirtualGrid'
-import { getStationAttackBonus } from '@utils/getAttackBonus'
+import { getStationDamageBoost } from '@utils/getAttackBonus'
 import { CopyCoords } from '@components/popups/Coords'
 import { PokeMove } from '@components/popups/PokeMove'
 
@@ -58,7 +56,8 @@ export function StationPopup(station) {
       )}
       <StationMedia {...station} />
       {station.battle_start < Date.now() / 1000 &&
-        station.battle_end > Date.now() / 1000 && (
+        station.battle_end > Date.now() / 1000 &&
+        !!station.total_stationed_pokemon && (
           <ExpandCollapse>
             <StationAttackBonus {...station} />
             <ExpandWithState
@@ -351,19 +350,14 @@ function StationAttackBonus({ total_stationed_pokemon, total_stationed_gmax }) {
   const { t } = useTranslation()
   return (
     <Stack alignItems="center">
-      <Rating
-        value={getStationAttackBonus(total_stationed_pokemon)}
-        readOnly
-        icon={<LockOpenIcon fontSize="inherit" />}
-        emptyIcon={<LockIcon fontSize="inherit" />}
-        max={4}
-      />
       <Typography variant="caption">
-        {t('battle_bonus')} &nbsp;(
+        {t('battle_bonus')}: +{getStationDamageBoost(total_stationed_pokemon)}%
+        <br />
+        {t('placed_pokemon')}:{' '}
         {total_stationed_gmax === undefined || total_stationed_gmax === null
           ? ''
-          : `${total_stationed_gmax} / `}
-        {total_stationed_pokemon} / 40)
+          : `${total_stationed_gmax}/`}
+        {total_stationed_pokemon}/40
       </Typography>
     </Stack>
   )
@@ -389,15 +383,11 @@ function StationContent({ start_time, end_time, id }) {
 /** @param {import('@rm/types').Station} props */
 function StationMons({ id }) {
   const { t: tId } = useTranslateById()
-  const { t } = useTranslation()
   const mons = useGetStationMons(id)
   const icons = useMemory((s) => s.Icons)
 
   return (
     <CardContent sx={{ my: 1, p: 0, height: 130 }}>
-      <Typography variant="h6" align="center">
-        {t('placed_pokemon')}
-      </Typography>
       <VirtualGrid data={mons} xs={6}>
         {(index, mon) => {
           const caption = tId(`${mon.pokemon_id}-${mon.form}`)