+++ src/services/queries/available.js
@@ -10,6 +10,7 @@ export const GET_MAP_DATA = gql`
       questConditions
       icons
       audio
+      supportsShinyStats
     }
   }
 `