+++ src/services/queries/pokemon.js
@@ -16,7 +16,6 @@ const core = gql`
     expire_timestamp_verified
     updated
     display_pokemon_id
-    ditto_form
     seen_type
   }
 `
@@ -133,6 +132,16 @@ export const GET_IVS_PVP = gql`
       ...CorePokemon
       ...Iv
       ...Pvp
+    }
+  }
+`
+
+export const GET_POKEMON_SHINY_STATS = gql`
+  query PokemonShinyStats($pokemonId: Int!, $form: Int) {
+    pokemonShinyStats(pokemon_id: $pokemonId, form: $form) {
+      shiny_seen
+      encounters_seen
+      since_date
     }
   }
 `